pyproject.toml
[tool.poetry]
name = "webnetlab"
version = "0.1.0"
description = "Network labs with web-interface using containerized NOSes"
authors = ["Grigoriy Macheev <gmacheev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
Jinja2 = "^3.1.2"
PyYAML = "^6.0"
python-dotenv = "^0.20.0"
SQLAlchemy = "^1.4.40"
alembic = "^1.8.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
